Add punnycode / IDN support to the domain name.
Provide a method for Domain and Vhost model which convert domains or vhosts to punnycode. This method is also used for the vhost.conf template which is created by the backend service. Database store domain names as punnycode but the output will always be unicode including umlauts or other special chars.