Workaround for Django "MySQL server has gone away"
If, for some reason, the MySQL / Percona server is not available for X-amount of time a timeout happen and the backend service doesn't try to reconnect to the service. By closing the connection the persisten open connection is closed and open again for the service. This is required on all backend functions.

Close the connection with from django.db import connection; connection.close() when you know that your program is going to be idle for a long time.

References:
- https://code.djangoproject.com/ticket/21597
- https://code.djangoproject.com/ticket/21597#comment:29