Provide TextField for redirect addresses to support more than 255 chars
Because 255 chars are really limited for a list of mail address destinations we should migrate to TextField. Anyway we should limit the display of these addresses in the UI. Additional to that a user could insert the list in formats like "new line seperated", "space seperated", etc. we should convert this during the export - so the user could deal with his own input (if prefered newlines).