Fix umask for directories create by Kumquat backend
To provide a secure backend service we set the umask for the socket. But
this option result in an umask for all folders and files created by the
backend service.

For example:

  770 for htdocs folder
  770 for logs folder

Both owned by www user and group. So the Kumquat web service isn't able
to access the error.log and display it correclty in the UI.